+++ GaluhadPlugins/Pets/Strings/German/Pets.lua
@@ -522,9 +522,10 @@ _PETSTRINGS[439] = {[1]="Schwarze Kobra";[2]="In Dun Shûma gefunden, der Festun
 _PETSTRINGS[440] = {[1]="Bärenjunges des Bäckervergnügens";[2]="Durch Tauschhandel während des Jubiläumsfests erhältlich"}
 _PETSTRINGS[441] = {[1]="Rostbrauner Corgi";[2]="";}
 _PETSTRINGS[442] = {[1]="Immergrüne Sichelfliege";[2]=""}
-_PETSTRINGS[443] = {[1]="Gefleckte Schlange";[2]=""}
+_PETSTRINGS[443] = {[1]="Gefleckte Schlange";[2]="In Dun Shûma gefunden, der Festung des Königs."}
 _PETSTRINGS[444] = {[1]="Gestreifte Springmaus";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 14"}
 _PETSTRINGS[445] = {[1]="Gefleckte Jadeschlange";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 14"}
+_PETSTRINGS[446] = {[1]="Fähiger Eisufer-Fischer";[2]="Prachtstücke des Glanzes nur für begrenzte Zeit und in Abenteurer- und Reisenden-Schatzkästchen verfügbar."}
 
 _PETSTRINGS[500] = {[1]="Gestreifte Fledermaus";[2]=""}
 _PETSTRINGS[501] = {[1]="Hellbraune Fledermaus";[2]=""}